Call CreateRepl() to start interpreting
Instead of our maps and everything always being defined, use a
function to explicitly create a repl. Also pre-populate the repl
with the prompt and put us in insert mode just like a terminal repl.

Uses my Scratch plugin if available.

Move ValidateLanguage into CreateRepl since we only need to validate
when we open the repl.